chore: complete phase 0 and phase 1 remediation

+++ barbing-saloon-api/app/Models/BarberService.php
@@ -19,7 +19,7 @@ class BarberService extends Model
     ];
 
     protected $casts = [
-        'custom_price' => 'float',
+        'custom_price' => 'integer',
         'custom_duration' => 'integer',
         'is_offered' => 'boolean',
     ];
